+++ cedar-policy/src/frontend/validate.rs
@@ -283,7 +283,7 @@ mod test {
 "#.to_string();
 
         let result = json_validate(&call_json);
-        assert_validates_with_errors(result, 4);
+        assert_validates_with_errors(result, 2);
     }
 
     #[test]
@@ -383,7 +383,7 @@ mod test {
 "#.to_string();
 
         let result = json_validate(&call_json);
-        assert_validates_with_errors(result, 2);
+        assert_validates_with_errors(result, 1);
     }
 
     #[test]